chore: move bot difficulty
