+++ packages/maumau-server/src/bot/bot-difficulty.ts
